assuming internal units are mm, convert idf files which are in thou (mils)

+++ src/idfdata.h
@@ -18,7 +18,7 @@ class IdfData {
 		//support functions
 		std::string getstring(std::string &line);
 		int getint(std::string &line);
-		float getfloat(std::string &line);
+		double getfloat(std::string &line);
 		bool process_header(std::ifstream &stream);
 		bool process_board_outline(std::ifstream &stream);
 		bool process_unknown(std::ifstream &stream, std::string type);
